coordinates now float + fixed movement

+++ entity_managment/Entity/Entity.cpp
@@ -5,7 +5,7 @@ Entity::Entity() {
     coordinates.y = 200;
 }
 
-void Entity::updateCurrent(const sf::Vector2i _coordinates) {
+void Entity::updateCurrent(const sf::Vector2f _coordinates) {
     coordinates.x = _coordinates.x;
     coordinates.y = _coordinates.y;
 }
@@ -14,6 +14,6 @@ void Entity::updateCurrent(sf::Time dt) {
     //will be overriden
 }
 
-Entity::Entity(sf::Vector2i _coordinates) {
+Entity::Entity(sf::Vector2f _coordinates) {
     coordinates = _coordinates;
 }
